Optimized query execution using fetch type, fetch mode and query annotation

+++ ProductService/src/main/java/com/ganesh/productservice/DTO/ProductResponseDTO.java
@@ -13,7 +13,7 @@ public class ProductResponseDTO {
     UUID uuid;
     String name;
     String description;
-    Integer rating;
+    Double rating;
     Double price;
     String category;
     String currency;
